Api auth

+++ app/api/DeleteFleetSettingsById/route.tsx
@@ -1,7 +1,8 @@
 import prisma from 'components/prisma'
 import { NextRequest, NextResponse } from "next/server";
-
 import assert from 'assert';
+import { AVAILABLE_PERMISSIONS } from 'components/helpers/users';
+import { isRequestAuthorised } from 'components/helpers/auth';
 
 async function findFleet(fleetSettingsId: any) {
     var result = await prisma.fleetSettings.findFirst({
@@ -27,18 +28,21 @@ export async function POST(request: NextRequest) {
         assert.notStrictEqual(undefined, req.fleetSettingsId, 'fleetSettingsId required');
 
     } catch (bodyError) {
-        return NextResponse.json({ error: true, message: "information missing" });
-        return;
+        return NextResponse.json({ error: "information missing" }, { status: 400 });
     }
 
     var fleetSettingsId = req.fleetSettingsId
 
+    let authorised = await isRequestAuthorised(request.cookies, AVAILABLE_PERMISSIONS.editFleets, fleetSettingsId, "fleetsettings")
+    if (!authorised) {
+        return NextResponse.json({ error: "not authorized" }, { status: 401 });
+    }
+
     var fleet = await findFleet(fleetSettingsId)
     if (fleet) {
         await deleteFleet(fleetSettingsId)
-        return NextResponse.json({ error: false, fleet: fleet });
-    } else {
-        // User exists
-        return NextResponse.json({ error: true, message: 'fleet not found' });
+        return NextResponse.json({ res: fleet }, { status: 200 });
     }
+    return NextResponse.json({ error: 'fleet not found' }, { status: 400 });
+
 }
